Fix cache modularization split visibility
