+++ crates/rginx-http/src/cache/mod.rs
@@ -18,8 +18,11 @@ use crate::handler::{HttpBody, HttpResponse};
 
 mod entry;
 mod load;
+mod lookup;
+mod manager;
 mod policy;
 mod request;
+mod runtime;
 mod store;
 
 use entry::{
@@ -33,6 +36,9 @@ use policy::{header_value, request_requires_revalidation};
 #[cfg(test)]
 use policy::{response_is_storable, response_ttl};
 use request::{cache_request_bypass, render_cache_key};
+use runtime::PurgeSelector;
+pub(crate) use runtime::with_cache_status;
+use runtime::{build_conditional_headers, remove_cache_files_if_unindexed};
 use store::{
     CacheStoreError, cleanup_inactive_entries_in_zone, lock_index, purge_zone_entries,
     refresh_not_modified_response, remove_index_entry, store_response,
@@ -197,593 +203,6 @@ enum LookupDecision {
 enum FillLockDecision {
     Acquired(CacheFillGuard),
     Wait(OwnedNotified),
-}
-
-impl CacheManager {
-    pub(crate) fn from_config_with_notifier(
-        config: &ConfigSnapshot,
-        change_notifier: Option<CacheChangeNotifier>,
-    ) -> Result<Self> {
-        let zones = config
-            .cache_zones
-            .iter()
-            .map(|(name, zone)| {
-                std::fs::create_dir_all(&zone.path).map_err(|error| {
-                    Error::Server(format!(
-                        "failed to create cache zone `{name}` directory `{}`: {error}",
-                        zone.path.display()
-                    ))
-                })?;
-                let index = load_index_from_disk(zone.as_ref()).map_err(|error| {
-                    Error::Server(format!(
-                        "failed to load cache zone `{name}` index from `{}`: {error}",
-                        zone.path.display()
-                    ))
-                })?;
-                Ok((
-                    name.clone(),
-                    Arc::new(CacheZoneRuntime {
-                        config: zone.clone(),
-                        index: Mutex::new(index),
-                        io_lock: AsyncMutex::new(()),
-                        fill_locks: Arc::new(Mutex::new(HashMap::new())),
-                        stats: CacheZoneStats::default(),
-                        change_notifier: change_notifier.clone(),
-                    }),
-                ))
-            })
-            .collect::<Result<HashMap<_, _>>>()?;
-
-        Ok(Self { zones: Arc::new(zones) })
-    }
-
-    pub(crate) async fn lookup(
-        &self,
-        request: CacheRequest,
-        downstream_scheme: &str,
-        policy: &RouteCachePolicy,
-    ) -> CacheLookup {
-        let Some(zone) = self.zones.get(&policy.zone).cloned() else {
-            tracing::warn!(
-                zone = %policy.zone,
-                "cache policy references unknown zone; bypassing cache"
-            );
-            return CacheLookup::Bypass(CacheStatus::Bypass);
-        };
-
-        if cache_request_bypass(&request, policy) {
-            zone.record_bypass();
-            return CacheLookup::Bypass(CacheStatus::Bypass);
-        }
-
-        let key = render_cache_key(
-            &request.method,
-            &request.uri,
-            &request.headers,
-            downstream_scheme,
-            policy,
-        );
-        let request_forces_revalidation = request_requires_revalidation(&request.headers);
-        let read_cached_body = request.method != Method::HEAD;
-
-        loop {
-            let now = unix_time_ms(SystemTime::now());
-            match self.lookup_decision(&zone, &key, now, request_forces_revalidation) {
-                LookupDecision::FreshHit(entry) => {
-                    let cached_response = {
-                        let _io_guard = zone.io_lock.lock().await;
-                        read_cached_response(&zone, &key, &entry, read_cached_body).await
-                    };
-                    match cached_response {
-                        Ok(mut response) => {
-                            zone.record_hit();
-                            response
-                                .headers_mut()
-                                .insert(CACHE_STATUS_HEADER, CacheStatus::Hit.as_header_value());
-                            return CacheLookup::Hit(response);
-                        }
-                        Err(error) => {
-                            tracing::warn!(
-                                zone = %zone.config.name,
-                                key_hash = %entry.hash,
-                                %error,
-                                "failed to read cached response; treating as miss"
-                            );
-                            remove_index_entry(&zone, &key);
-                            remove_cache_files_if_unindexed(&zone, &key, &entry.hash).await;
-                        }
-                    }
-                }
-                LookupDecision::StaleWhileRevalidate(entry) => {
-                    match self
-                        .stale_response_from_entry(&zone, &key, &entry, read_cached_body)
-                        .await
-                    {
-                        Some(response) => return CacheLookup::Hit(response),
-                        None => continue,
-                    }
-                }
-                LookupDecision::Wait(waiter) => {
-                    waiter.await;
-                }
-                LookupDecision::Miss {
-                    cached_entry,
-                    fill_guard,
-                    cache_status,
-                    allow_stale_on_error,
-                } => {
-                    let (cached_metadata, conditional_headers) = if let Some(entry) = &cached_entry
-                    {
-                        match self.load_lookup_metadata(&zone, &key, entry).await {
-                            Some((metadata, conditional_headers)) => {
-                                (Some(metadata), conditional_headers)
-                            }
-                            None => {
-                                drop(fill_guard);
-                                continue;
-                            }
-                        }
-                    } else {
-                        (None, None)
-                    };
-                    if cache_status == CacheStatus::Miss {
-                        zone.record_miss();
-                    } else if cache_status == CacheStatus::Expired {
-                        zone.record_expired();
-                    }
-
-                    return CacheLookup::Miss(Box::new(CacheStoreContext {
-                        zone,
-                        policy: policy.clone(),
-                        key,
-                        cache_status,
-                        store_response: request.method == Method::GET,
-                        _fill_guard: Some(fill_guard),
-                        cached_entry,
-                        cached_metadata,
-                        allow_stale_on_error,
-                        revalidating: cache_status == CacheStatus::Revalidated,
-                        conditional_headers,
-                        read_cached_body,
-                    }));
-                }
-            }
-        }
-    }
-
-    pub(crate) async fn store_response(
-        &self,
-        context: CacheStoreContext,
-        response: HttpResponse,
-    ) -> HttpResponse {
-        let status = context.cache_status;
-        let response = if context.store_response {
-            match store_response(context, response).await {
-                Ok(response) => response,
-                Err(error) => {
-                    tracing::warn!(%error, "failed to store cached response");
-                    crate::handler::text_response(
-                        StatusCode::BAD_GATEWAY,
-                        "text/plain; charset=utf-8",
-                        format!("failed to read upstream response while caching: {error}\n"),
-                    )
-                }
-            }
-        } else {
-            response
-        };
-
-        with_cache_status(response, status)
-    }
-
-    pub(crate) async fn complete_not_modified(
-        &self,
-        context: CacheStoreContext,
-        response: HttpResponse,
-    ) -> std::result::Result<HttpResponse, CacheStoreError> {
-        refresh_not_modified_response(context, response).await
-    }
-
-    pub(crate) fn snapshot(&self) -> Vec<CacheZoneRuntimeSnapshot> {
-        let mut snapshots = self.zones.values().map(|zone| zone.snapshot()).collect::<Vec<_>>();
-        snapshots.sort_by(|left, right| left.zone_name.cmp(&right.zone_name));
-        snapshots
-    }
-
-    pub(crate) async fn cleanup_inactive_entries(&self) {
-        for zone in self.zones.values() {
-            cleanup_inactive_entries_in_zone(zone).await;
-        }
-    }
-
-    pub(crate) async fn purge_zone(
-        &self,
-        zone_name: &str,
-    ) -> std::result::Result<CachePurgeResult, String> {
-        let zone = self
-            .zones
-            .get(zone_name)
-            .cloned()
-            .ok_or_else(|| format!("unknown cache zone `{zone_name}`"))?;
-        Ok(purge_zone_entries(zone, PurgeSelector::All).await)
-    }
-
-    pub(crate) async fn purge_key(
-        &self,
-        zone_name: &str,
-        key: &str,
-    ) -> std::result::Result<CachePurgeResult, String> {
-        let zone = self
-            .zones
-            .get(zone_name)
-            .cloned()
-            .ok_or_else(|| format!("unknown cache zone `{zone_name}`"))?;
-        Ok(purge_zone_entries(zone, PurgeSelector::Exact(key.to_string())).await)
-    }
-
-    pub(crate) async fn purge_prefix(
-        &self,
-        zone_name: &str,
-        prefix: &str,
-    ) -> std::result::Result<CachePurgeResult, String> {
-        let zone = self
-            .zones
-            .get(zone_name)
-            .cloned()
-            .ok_or_else(|| format!("unknown cache zone `{zone_name}`"))?;
-        Ok(purge_zone_entries(zone, PurgeSelector::Prefix(prefix.to_string())).await)
-    }
-
-    fn lookup_decision(
-        &self,
-        zone: &Arc<CacheZoneRuntime>,
-        key: &str,
-        now: u64,
-        request_forces_revalidation: bool,
-    ) -> LookupDecision {
-        let mut index = lock_index(&zone.index);
-        match index.entries.get_mut(key) {
-            Some(entry)
-                if now <= entry.expires_at_unix_ms
-                    && !entry.must_revalidate
-                    && !request_forces_revalidation =>
-            {
-                entry.last_access_unix_ms = now;
-                LookupDecision::FreshHit(entry.clone())
-            }
-            Some(entry) => {
-                entry.last_access_unix_ms = now;
-                match zone.fill_lock_decision(key) {
-                    FillLockDecision::Acquired(fill_guard) => {
-                        let expired = now > entry.expires_at_unix_ms;
-                        let cache_status =
-                            if expired { CacheStatus::Expired } else { CacheStatus::Revalidated };
-                        let allow_stale_on_error = expired
-                            && entry.stale_if_error_until_unix_ms.is_some_and(|until| now <= until);
-                        LookupDecision::Miss {
-                            cached_entry: Some(entry.clone()),
-                            fill_guard,
-                            cache_status,
-                            allow_stale_on_error,
-                        }
-                    }
-                    FillLockDecision::Wait(_waiter)
-                        if now > entry.expires_at_unix_ms
-                            && entry
-                                .stale_while_revalidate_until_unix_ms
-                                .is_some_and(|until| now <= until) =>
-                    {
-                        LookupDecision::StaleWhileRevalidate(entry.clone())
-                    }
-                    FillLockDecision::Wait(waiter) => LookupDecision::Wait(waiter),
-                }
-            }
-            None => match zone.fill_lock_decision(key) {
-                FillLockDecision::Acquired(fill_guard) => LookupDecision::Miss {
-                    cached_entry: None,
-                    fill_guard,
-                    cache_status: CacheStatus::Miss,
-                    allow_stale_on_error: false,
-                },
-                FillLockDecision::Wait(waiter) => LookupDecision::Wait(waiter),
-            },
-        }
-    }
-
-    async fn load_lookup_metadata(
-        &self,
-        zone: &Arc<CacheZoneRuntime>,
-        key: &str,
-        entry: &CacheIndexEntry,
-    ) -> Option<(CacheMetadata, Option<CacheConditionalHeaders>)> {
-        let metadata = {
-            let _io_guard = zone.io_lock.lock().await;
-            let paths = cache_paths(&zone.config.path, &entry.hash);
-            read_cache_metadata(&paths.metadata).await
-        };
-        let metadata = match metadata {
-            Ok(metadata) => metadata,
-            Err(error) => {
-                tracing::warn!(
-                    zone = %zone.config.name,
-                    key_hash = %entry.hash,
-                    %error,
-                    "failed to read cache metadata; removing entry"
-                );
-                remove_index_entry(zone, key);
-                remove_cache_files_if_unindexed(zone, key, &entry.hash).await;
-                return None;
-            }
-        };
-        if metadata.key != key {
-            tracing::warn!(
-                zone = %zone.config.name,
-                key = %key,
-                cached_key = %metadata.key,
-                key_hash = %entry.hash,
-                "cache metadata key mismatch; removing entry"
-            );
-            remove_index_entry(zone, key);
-            remove_cache_files_if_unindexed(zone, key, &entry.hash).await;
-            return None;
-        }
-        let headers = match metadata.headers_map() {
-            Ok(headers) => headers,
-            Err(error) => {
-                tracing::warn!(
-                    zone = %zone.config.name,
-                    key_hash = %entry.hash,
-                    %error,
-                    "failed to decode cached response headers; removing entry"
-                );
-                remove_index_entry(zone, key);
-                remove_cache_files_if_unindexed(zone, key, &entry.hash).await;
-                return None;
-            }
-        };
-        let conditional_headers = build_conditional_headers(&headers);
-        Some((metadata, conditional_headers))
-    }
-
-    async fn stale_response_from_entry(
-        &self,
-        zone: &Arc<CacheZoneRuntime>,
-        key: &str,
-        entry: &CacheIndexEntry,
-        read_cached_body: bool,
-    ) -> Option<HttpResponse> {
-        let (metadata, response) = {
-            let _io_guard = zone.io_lock.lock().await;
-            let paths = cache_paths(&zone.config.path, &entry.hash);
-            let metadata = read_cache_metadata(&paths.metadata).await.ok()?;
-            let response =
-                build_cached_response(&paths.body, &metadata, read_cached_body).await.ok()?;
-            (metadata, response)
-        };
-        if metadata.key != key {
-            tracing::warn!(
-                zone = %zone.config.name,
-                key = %key,
-                cached_key = %metadata.key,
-                key_hash = %entry.hash,
-                "cache metadata key mismatch while serving stale entry; removing entry"
-            );
-            remove_index_entry(zone, key);
-            remove_cache_files_if_unindexed(zone, key, &entry.hash).await;
-            return None;
-        }
-        zone.record_stale();
-        Some(with_cache_status(response, CacheStatus::Stale))
-    }
-}
-
-impl CacheRequest {
-    pub(crate) fn from_request(request: &Request<HttpBody>) -> Self {
-        Self {
-            method: request.method().clone(),
-            uri: request.uri().clone(),
-            headers: request.headers().clone(),
-        }
-    }
-}
-
-impl CacheStoreContext {
-    pub(crate) fn cache_status(&self) -> CacheStatus {
-        self.cache_status
-    }
-
-    pub(crate) fn apply_conditional_request_headers(&self, headers: &mut HeaderMap) {
-        let Some(conditional_headers) = &self.conditional_headers else {
-            return;
-        };
-        if let Some(value) = conditional_headers.if_none_match.clone() {
-            headers.insert(IF_NONE_MATCH, value);
-        }
-        if let Some(value) = conditional_headers.if_modified_since.clone() {
-            headers.insert(IF_MODIFIED_SINCE, value);
-        }
-    }
-
-    pub(crate) fn should_refresh_from_not_modified(&self, status: StatusCode) -> bool {
-        self.cached_entry.is_some() && status == StatusCode::NOT_MODIFIED
-    }
-
-    pub(crate) fn can_serve_stale_on_error(&self) -> bool {
-        self.allow_stale_on_error && self.cached_entry.is_some() && self.cached_metadata.is_some()
-    }
-
-    pub(crate) async fn serve_stale_on_error(&self) -> Option<HttpResponse> {
-        let Some(entry) = &self.cached_entry else {
-            return None;
-        };
-        let Some(metadata) = &self.cached_metadata else {
-            return None;
-        };
-        let response = {
-            let _io_guard = self.zone.io_lock.lock().await;
-            let paths = cache_paths(&self.zone.config.path, &entry.hash);
-            build_cached_response(&paths.body, metadata, self.read_cached_body).await
-        };
-        match response {
-            Ok(response) => {
-                self.zone.record_stale();
-                Some(with_cache_status(response, CacheStatus::Stale))
-            }
-            Err(error) => {
-                tracing::warn!(
-                    zone = %self.zone.config.name,
-                    key = %self.key,
-                    %error,
-                    "failed to serve stale cache entry"
-                );
-                remove_index_entry(&self.zone, &self.key);
-                remove_cache_files_if_unindexed(&self.zone, &self.key, &entry.hash).await;
-                None
-            }
-        }
-    }
-}
-
-impl CacheZoneRuntime {
-    fn snapshot(&self) -> CacheZoneRuntimeSnapshot {
-        let index = lock_index(&self.index);
-        CacheZoneRuntimeSnapshot {
-            zone_name: self.config.name.clone(),
-            path: self.config.path.clone(),
-            max_size_bytes: self.config.max_size_bytes,
-            inactive_secs: self.config.inactive.as_secs(),
-            default_ttl_secs: self.config.default_ttl.as_secs(),
-            max_entry_bytes: self.config.max_entry_bytes,
-            entry_count: index.entries.len(),
-            current_size_bytes: index.current_size_bytes,
-            hit_total: self.stats.hit_total.load(Ordering::Relaxed),
-            miss_total: self.stats.miss_total.load(Ordering::Relaxed),
-            bypass_total: self.stats.bypass_total.load(Ordering::Relaxed),
-            expired_total: self.stats.expired_total.load(Ordering::Relaxed),
-            stale_total: self.stats.stale_total.load(Ordering::Relaxed),
-            revalidated_total: self.stats.revalidated_total.load(Ordering::Relaxed),
-            write_success_total: self.stats.write_success_total.load(Ordering::Relaxed),
-            write_error_total: self.stats.write_error_total.load(Ordering::Relaxed),
-            eviction_total: self.stats.eviction_total.load(Ordering::Relaxed),
-            purge_total: self.stats.purge_total.load(Ordering::Relaxed),
-            inactive_cleanup_total: self.stats.inactive_cleanup_total.load(Ordering::Relaxed),
-        }
-    }
-
-    fn fill_lock_decision(self: &Arc<Self>, key: &str) -> FillLockDecision {
-        let mut fill_locks =
-            self.fill_locks.lock().unwrap_or_else(|poisoned| poisoned.into_inner());
-        if let Some(notify) = fill_locks.get(key).cloned() {
-            return FillLockDecision::Wait(notify.notified_owned());
-        }
-        let notify = Arc::new(Notify::new());
-        fill_locks.insert(key.to_string(), notify.clone());
-        FillLockDecision::Acquired(CacheFillGuard {
-            key: key.to_string(),
-            fill_locks: Arc::downgrade(&self.fill_locks),
-            notify,
-        })
-    }
-
-    fn record_hit(&self) {
-        self.record_counter(&self.stats.hit_total, 1);
-    }
-
-    fn record_miss(&self) {
-        self.record_counter(&self.stats.miss_total, 1);
-    }
-
-    fn record_bypass(&self) {
-        self.record_counter(&self.stats.bypass_total, 1);
-    }
-
-    fn record_expired(&self) {
-        self.record_counter(&self.stats.expired_total, 1);
-    }
-
-    fn record_stale(&self) {
-        self.record_counter(&self.stats.stale_total, 1);
-    }
-
-    pub(super) fn record_revalidated(&self) {
-        self.record_counter(&self.stats.revalidated_total, 1);
-    }
-
-    pub(super) fn record_write_success(&self) {
-        self.record_counter(&self.stats.write_success_total, 1);
-    }
-
-    pub(super) fn record_write_error(&self) {
-        self.record_counter(&self.stats.write_error_total, 1);
-    }
-
-    pub(super) fn record_evictions(&self, count: usize) {
-        if count > 0 {
-            self.record_counter(&self.stats.eviction_total, count as u64);
-        }
-    }
-
-    pub(super) fn record_purge(&self, count: usize) {
-        if count > 0 {
-            self.record_counter(&self.stats.purge_total, count as u64);
-        }
-    }
-
-    pub(super) fn record_inactive_cleanup(&self, count: usize) {
-        if count > 0 {
-            self.record_counter(&self.stats.inactive_cleanup_total, count as u64);
-        }
-    }
-
-    fn record_counter(&self, counter: &AtomicU64, value: u64) {
-        counter.fetch_add(value, Ordering::Relaxed);
-    }
-
-    pub(super) fn notify_changed(&self) {
-        if let Some(notifier) = &self.change_notifier {
-            notifier(&self.config.name);
-        }
-    }
-}
-
-impl Drop for CacheFillGuard {
-    fn drop(&mut self) {
-        if let Some(fill_locks) = self.fill_locks.upgrade() {
-            fill_locks.lock().unwrap_or_else(|poisoned| poisoned.into_inner()).remove(&self.key);
-        }
-        self.notify.notify_waiters();
-    }
-}
-
-pub(crate) fn with_cache_status(mut response: HttpResponse, status: CacheStatus) -> HttpResponse {
-    response.headers_mut().insert(CACHE_STATUS_HEADER, status.as_header_value());
-    response
-}
-
-async fn remove_cache_files_if_unindexed(zone: &CacheZoneRuntime, key: &str, hash: &str) {
-    let _io_guard = zone.io_lock.lock().await;
-    if lock_index(&zone.index).entries.contains_key(key) {
-        return;
-    }
-    let paths = cache_paths(&zone.config.path, hash);
-    let _ = fs::remove_file(paths.metadata).await;
-    let _ = fs::remove_file(paths.body).await;
-}
-
-fn build_conditional_headers(headers: &HeaderMap) -> Option<CacheConditionalHeaders> {
-    let if_none_match =
-        header_value(headers, ETAG).and_then(|value| HeaderValue::from_str(&value).ok());
-    let if_modified_since =
-        header_value(headers, LAST_MODIFIED).and_then(|value| HeaderValue::from_str(&value).ok());
-    (if_none_match.is_some() || if_modified_since.is_some())
-        .then_some(CacheConditionalHeaders { if_none_match, if_modified_since })
-}
-
-#[derive(Debug, Clone)]
-pub(super) enum PurgeSelector {
-    All,
-    Exact(String),
-    Prefix(String),
 }
 
 #[cfg(test)]
